+++ package.json
@@ -38,7 +38,7 @@
 		"@storybook/test": "^8.4.7",
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "^2.9.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@types/bun": "^1.1.14",
 		"autoprefixer": "^10.4.20",
 		"daisyui": "^4.12.22",
@@ -61,6 +61,7 @@
 	},
 	"dependencies": {
 		"@flaticon/flaticon-uicons": "^3.3.1",
+		"@inlang/paraglide-sveltekit": "^0.15.0",
 		"@lucia-auth/adapter-prisma": "^4.0.1",
 		"@pothos/core": "^4.3.0",
 		"@pothos/plugin-prisma": "^4.4.0",
