+++ src/routes/+layout.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+	import { i18n } from '$lib/i18n';
+	import { ParaglideJS } from '@inlang/paraglide-sveltekit';
 	import '../app.css';
 	let { children } = $props();
 </script>
@@ -7,8 +9,10 @@
 	{@render children()}
 </div>
 
-<style>
-	.layout {
-		@apply h-screen w-screen bg-base-100;
-	}
-</style>
+<ParaglideJS {i18n}>
+	<style>
+		.layout {
+			@apply h-screen w-screen bg-base-100;
+		}
+	</style>
+</ParaglideJS>
