+++ src/routes/+page.server.ts
@@ -1,18 +1,5 @@
-import { prisma } from '$lib/server/prisma';
-import { redirect } from '@sveltejs/kit';
+import { validateSession } from '$lib/server/auth';
 
 export async function load(event) {
-	const sessionId = event.cookies.get('auth_session');
-	if (!sessionId) {
-		redirect(303, '/login');
-	}
-	const user = await prisma.session.findUnique({
-		where: {
-			id: sessionId,
-		},
-	});
-	if (!user) {
-		redirect(300, '/login');
-	}
-	return {};
+	await validateSession(event);
 }
