+++ vite.config.ts
@@ -1,8 +1,15 @@
 import { defineConfig } from 'vitest/config';
+import { paraglide } from '@inlang/paraglide-sveltekit/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 
 export default defineConfig({
-	plugins: [sveltekit()],
+	plugins: [
+		sveltekit(),
+		paraglide({
+			project: './project.inlang',
+			outdir: './src/lib/paraglide',
+		}),
+	],
 
 	test: {
 		include: ['src/**/*.{test,spec}.{js,ts}'],
